requirements.txt
pytest~=7.2.2
pytest-xdist
pytest-ordering
pytest-html
allure-pytest
pytest-rerunfailures
cryptography
pyOpenSSL
certifi
selenium~=4.8.2
document
openpyxl
pytest-xlsx